# packages/nextjs/lambda/enhanced-openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Marketplace Shopping Agent Functions",
    "description": "Enhanced function definitions for autonomous shopping operations with comprehensive parameter validation and response schemas",
    "version": "2.0.0"
  },
  "paths": {
    "/searchProducts": {
      "post": {
        "summary": "Search for products in the marketplace",
        "description": "Search through all 21+ verified sustainable products using natural language queries with advanced filtering capabilities",
        "operationId": "searchProducts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Natural language search query for products (e.g., 'sustainable electronics', 'bamboo products', 'AI-powered devices')",
                    "minLength": 1,
                    "maxLength": 200,
                    "examples": ["sustainable electronics", "bamboo laptop stand", "AI-powered smart watch"]
                  },
                  "category": {
                    "type": "string",
                    "description": "Product category filter",
                    "enum": ["Electronics", "Clothing", "Digital", "Sports", "Books", "Home & Garden", "Beauty", "Automotive"],
                    "examples": ["Electronics", "Clothing"]
                  },
                  "maxPrice": {
                    "type": "number",
                    "description": "Maximum price in USD",
                    "minimum": 0,
                    "maximum": 10000,
                    "examples": [50, 100, 500]
                  },
                  "sustainabilityMin": {
                    "type": "number",
                    "description": "Minimum sustainability score (0-100 scale, recommend 80+ for eco-conscious buyers)",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 80,
                    "examples": [80, 90, 95]
                  }
                },
                "required": ["query"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "number" },
                              "name": { "type": "string" },
                              "description": { "type": "string" },
                              "category": { "type": "string" },
                              "price": { "type": "string" },
                              "priceUSD": { "type": "number" },
                              "sustainabilityScore": { "type": "number" },
                              "averageRating": { "type": "number" },
                              "certifications": { "type": "array", "items": { "type": "string" } }
                            }
                          }
                        },
                        "totalFound": { "type": "number" },
                        "totalAvailable": { "type": "number" },
                        "source": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}